REFACTOR: Moved functions related to file system into file-system folder
