+++ actions/file-system/getImagesFromFileSystem.ts
